requirements.txt
Pyrogram==1.4.8
TgCrypto==1.2.3
loguru==0.6.0
python-dotenv==0.19.2